Ambient in scena, Makefile fichi

+++ src/ft_operation_mouse.c
@@ -6,8 +6,8 @@ int		mouse_press(int button, int x, int y, t_rtv *p)
 	{
 		if (button == 1)
 		{
-			p->camera->angle.y += asin((double)(x - p->x0) / p->width);
-			p->camera->angle.x += asin((double)(p->y0 - y) / p->width);
+			p->camera->angle.y += asin((double)(x - p->x0) / (double)p->width);
+			p->camera->angle.x += asin((double)(p->y0 - y) / (double)p->width);
 			ft_paint_scene(p);
 		}
 		if (button == 4 || button == 5)
@@ -29,8 +29,8 @@ int		mouse_movement(int x, int y, t_rtv *p)
 
 	if (y >= 0 && y <= (HIGHT - 1) && x >= 0 && x <= (WIDHT - 1))
 	{
-		p->camera->angle.y += asin((double)(x - p->x0) / p->width);
-		p->camera->angle.x += asin((double)(p->y0 - y) / p->width);
+		p->camera->angle.y += asin((double)(x - p->x0) / (double)p->width);
+		p->camera->angle.x += asin((double)(p->y0 - y) / (double)p->width);
 	}
 	ft_paint_scene(p);
 	return (0);
